dags-dev/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
idna==3.11
minio==7.2.20
pycparser==2.23
pycryptodome==3.23.0
python-dotenv==1.2.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.2